Split all instructions into multiple files
