+++ emulator/src/core/emulator.rs
@@ -372,140 +372,6 @@ impl Emulator {
         Ok(())
     }
 
-    fn mvi(&mut self, r: char) -> EResult<()> {
-        self.reg[r] = self.read_byte()?;
-        Ok(())
-    }
-
-    fn mvi_adr(&mut self) -> EResult<()> {
-        // Move byte 2 to address in HL
-        let byte = self.read_byte()?;
-        let adr = self.reg["hl"];
-        self.ram[adr] = byte;
-        Ok(())
-    }
-
-    fn resolve_mov(&mut self, opcode: u8) -> EResult<()> {
-        let registers = ['b', 'c', 'd', 'e', 'h', 'l', 'm', 'a'];
-        let opcode_rel = opcode - 0x40;
-        let dst_idx = opcode_rel >> 3;
-        let src_idx = opcode_rel - (dst_idx << 3);
-        if dst_idx == 6 {
-            self.ram[self.reg["hl"]] = self.reg[registers[src_idx as usize]];
-        } else {
-            if src_idx == 6 {
-                self.reg[registers[dst_idx as usize]] = self.ram[self.reg["hl"]];
-            } else {
-                self.mov(registers[dst_idx as usize], registers[src_idx as usize])?;
-            }
-        }
-        Ok(())
-    }
-
-    fn mov(&mut self, dst: char, src: char) -> EResult<()> {
-        self.reg[dst] = self.reg[src];
-        Ok(())
-    }
-
-    fn lxi(&mut self, dst: &str) -> EResult<()> {
-        self.reg[dst] = self.read_addr()?;
-        Ok(())
-    }
-
-    fn jmp_not(&mut self, flag: &str) -> EResult<()> {
-        if !self.reg.get_flag(flag) {
-            self.pc = self.read_addr()?;
-        } else {
-            self.pc += 2;
-        }
-        Ok(())
-    }
-
-    fn jmp_if(&mut self, flag: &str) -> EResult<()> {
-        if self.reg.get_flag(flag) {
-            self.pc = self.read_addr()?;
-        } else {
-            self.pc += 2;
-        }
-        Ok(())
-    }
-
-    fn call_not(&mut self, flag: &str) -> EResult<()> {
-        if !self.reg.get_flag(flag) {
-            self.call_imm()?;
-        } else {
-            self.pc += 2;
-        }
-        Ok(())
-    }
-
-    fn call_if(&mut self, flag: &str) -> EResult<()> {
-        if self.reg.get_flag(flag) {
-            self.call_imm()?;
-        } else {
-            self.pc += 2;
-        }
-        Ok(())
-    }
-
-    fn call_imm(&mut self) -> EResult<()> {
-        let adr = self.read_addr()?;
-        self.push(self.pc)?;
-        self.pc = adr;
-        Ok(())
-    }
-
-    fn call(&mut self, adr: u16) -> EResult<()> {
-        self.push(self.pc)?;
-        self.pc = adr;
-        Ok(())
-    }
-
-    fn ret_if(&mut self, flag: &str) -> EResult<()> {
-        if self.reg.get_flag(flag) {
-            self.ret()?;
-        }
-        Ok(())
-    }
-
-    fn ret_not(&mut self, flag: &str) -> EResult<()> {
-        if !self.reg.get_flag(flag) {
-            self.ret()?;
-        }
-        Ok(())
-    }
-
-    fn ret(&mut self) -> EResult<()> {
-        self.pc = self.pop()?;
-        Ok(())
-    }
-
-    fn push(&mut self, val: u16) -> EResult<()> {
-        if self.sp < 2 {
-            return Err("PUSH: No more stack space");
-        }
-        self.sp -= 1;
-        self.ram[self.sp] = (val >> 8) as u8;
-        self.sp -= 1;
-        self.ram[self.sp] = val as u8;
-        Ok(())
-    }
-
-    fn push_reg(&mut self, reg: &str) -> EResult<()> {
-        self.push(self.reg[reg])
-    }
-
-    fn pop(&mut self) -> EResult<u16> {
-        if self.sp + 2 > self.ram.size() as u16 {
-            return Err("POP: No return address on the stack");
-        }
-        let low = self.ram[self.sp] as u16;
-        self.sp += 1;
-        let high = self.ram[self.sp] as u16;
-        self.sp += 1;
-        Ok((high << 8) | low)
-    }
-
     fn read_byte(&mut self) -> EResult<u8> {
         if self.pc + 1 > self.ram.size() as u16 {
             return Err("READ_BYTE: Not enough bytes available");
@@ -527,222 +393,3 @@ impl Emulator {
 }
 
 mod instructions;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::ram::*;
-    use crate::kreator::assembler::Assembler;
-    use std::{
-        fs::*,
-        io::{self, Read},
-    };
-
-    fn load_asm_file(emulator: &mut Emulator, path: &str) -> io::Result<()> {
-        let mut file = File::open(path)?;
-        let mut buf = String::new();
-        file.read_to_string(&mut buf)?;
-        let asmblr = Assembler::new(&buf);
-        let mc = asmblr.assemble().expect("Fuck");
-        emulator.ram.load_vec(mc, 0);
-        Ok(())
-    }
-
-    #[test]
-    fn mvi() -> io::Result<()> {
-        let mut emu = Emulator::new();
-        load_asm_file(&mut emu, "./src/core/asm/mvi.s")?;
-
-        // Check MVI reg, D8
-        let regs = ['b', 'c', 'd', 'e', 'h', 'l', 'a'];
-        for i in 0..7 {
-            emu.execute_next().expect("Fuck");
-            assert_eq!(emu.reg[regs[i]], (0x1d + i) as u8);
-        }
-        emu.execute_next().expect("Fuck");
-
-        // Check MVI M, D8
-        assert_eq!(emu.ram[emu.reg["hl"]], 0x24);
-        Ok(())
-    }
-
-    #[test]
-    fn mov() -> io::Result<()> {
-        let mut emu = Emulator::new();
-        load_asm_file(&mut emu, "./src/core/asm/mov.s")?;
-        for _ in 0..8 {
-            emu.execute_next().expect("Fuck");
-        }
-        for i in 0..8 {
-            emu.execute_next().expect("Fuck");
-            assert_eq!(emu.reg['b'], (0x1d + i) as u8);
-        }
-
-        // Test MOV M, SRC
-        emu.execute_next().expect("Fuck");
-        assert_eq!(emu.ram[emu.reg["hl"]], emu.reg['b']);
-
-        // Test HLT
-        emu.execute_next().expect("Fuck");
-        assert_eq!(emu.running, false);
-        Ok(())
-    }
-
-    #[test]
-    fn lxi() -> io::Result<()> {
-        let mut emu = Emulator::new();
-        load_asm_file(&mut emu, "./src/core/asm/lxi.s")?;
-
-        let regs = ["bc", "de", "hl"];
-        for i in 1..4 {
-            emu.execute_next().expect("Fuck");
-            assert_eq!(emu.reg[regs[i-1]], (i * 256 + i + 4) as u16);
-        }
-        emu.execute_next().expect("Fuck");
-        assert_eq!(emu.sp, 0x0408);
-        Ok(())
-    }
-
-    #[test]
-    fn push_pop() {
-        let mut e = Emulator::new();
-
-        e.sp = 0x3fff;
-        e.push(0xabcd).expect("Push failed");
-        assert_eq!(e.sp, 0x3ffd);
-        assert_eq!(0xabcd, e.pop().expect("Fuck"));
-        assert_eq!(e.sp, 0x3fff);
-        assert_eq!(e.pop(), Err("POP: No return address on the stack"));
-
-        e.sp = 0x1;
-        assert_eq!(e.push(0x1234), Err("PUSH: No more stack space"));
-    }
-
-    #[test]
-    fn call_ret() {
-        let mut e = Emulator::new();
-
-        e.sp = 0x3fff;
-        e.ram[0x1234] = 0xc9;
-
-        e.call(0x1234).expect("Fuck");
-        assert_eq!(e.sp, 0x3fff - 2);
-        assert_eq!(e.pc, 0x1234);
-
-        e.execute_next().expect("Fuck");
-        assert_eq!(e.pc, 0x0);
-    }
-
-    #[test]
-    fn jmp_if() {
-        let mut e = Emulator::new();
-
-        e.ram.load_vec(vec![0x04, 0x00, 0x00, 0x00], 0);
-
-        // Performs
-        // a) one failing jmp (-> pc = 2)
-        // b) one succeeding jmp (pc = ram[pc] = ram[2] -> 0)
-        // c) Back in starting position
-        // -> Repeat for each flag
-        for flag in vec!["zero", "carry", "sign", "parity", "aux"] {
-            e.jmp_if(flag).expect("");
-            assert_eq!(e.pc, 2);
-            e.reg.set_flag(flag, true);
-            e.jmp_if(flag).expect("");
-            assert_eq!(e.pc, 0);
-        }
-    }
-
-    #[test]
-    fn jmp_not() {
-        let mut e = Emulator::new();
-
-        e.ram.load_vec(vec![0x04, 0x00, 0x00, 0x00], 0);
-        e.reg.set_flags(0xff);
-
-        // same as tests::jmp_if
-        for flag in vec!["zero", "carry", "sign", "parity", "aux"] {
-            e.jmp_not(flag).expect("");
-            assert_eq!(e.pc, 2);
-            e.reg.flip_flag(flag);
-            e.jmp_not(flag).expect("");
-            assert_eq!(e.pc, 0);
-        }
-    }
-
-    #[test]
-    fn call_if() {
-        let mut e = Emulator::new();
-
-        e.sp = 0x3fff;
-
-        e.ram.load_vec(vec![0x00, 0x00, 0x11, 0x11], 0);
-
-        for flag in vec!["zero", "carry", "sign", "parity", "aux"] {
-            e.call_if(flag).expect("");
-            assert_eq!(e.pc, 2);
-            e.ret_if(flag).expect("");
-            assert_eq!(e.pc, 2);
-            e.reg.set_flag(flag, true);
-            e.call_if(flag).expect("");
-            assert_eq!(e.pc, 0x1111);
-            e.ret_if(flag).expect("");
-            assert_eq!(e.pc, 4);
-            e.pc = 0;
-        }
-    }
-
-    #[test]
-    fn call_not() {
-        let mut e = Emulator::new();
-
-        e.sp = 0x3fff;
-
-        e.ram.load_vec(vec![0x00, 0x00, 0x11, 0x11], 0);
-        e.reg.set_flags(0xff);
-
-        for flag in vec!["zero", "carry", "sign", "parity", "aux"] {
-            e.call_not(flag).expect("");
-            assert_eq!(e.pc, 2);
-            e.ret_not(flag).expect("");
-            assert_eq!(e.pc, 2);
-            e.reg.flip_flag(flag);
-            e.call_not(flag).expect("");
-            assert_eq!(e.pc, 0x1111);
-            e.ret_not(flag).expect("");
-            assert_eq!(e.pc, 4);
-            e.pc = 0;
-        }
-    }
-
-    #[test]
-    fn call() {
-        let mut e = Emulator::new();
-
-        e.sp = 0x3fff;
-
-        assert_eq!(e.pc, 0x0);
-        e.call_if("carry").expect("Fuck");
-        assert_eq!(e.pc, 0x2);
-        e.reg.set_flag("carry", true);
-        e.ram.load_vec(vec![0x34, 0x12], 2);
-        e.call_if("carry").expect("Fuck");
-        assert_eq!(e.pc, 0x1234);
-    }
-
-    #[test]
-    fn rst() {
-        let mut e = Emulator::new();
-
-        e.pc = 0x1111;
-        e.sp = 0x3fff;
-
-        e.ram
-            .load_vec(vec![0xc7, 0xcf, 0xd7, 0xdf, 0xe7, 0xef, 0xf7, 0xff], e.pc);
-        for i in 0x1111..0x1119 {
-            e.pc = i as u16;
-            e.execute_next().expect("Fuck");
-            assert_eq!(e.pc, (i - 0x1111) * 8);
-        }
-    }
-}
