+++ emulator/src/core/emulator/instructions/mod.rs
@@ -1,2 +1,5 @@
 mod arithmetic;
 mod logic;
+mod r#move;
+mod controlflow;
+mod stack;
