+++ emulator/src/core/emulator/instructions/stack.rs
@@ -1,0 +1,53 @@
+use super::super::{Emulator, EResult};
+
+const REGISTERS: [char; 8] = ['b', 'c', 'd', 'e', 'h', 'l', 'm', 'a'];
+
+impl Emulator {
+    pub fn push(&mut self, val: u16) -> EResult<()> {
+        if self.sp < 2 {
+            return Err("PUSH: No more stack space");
+        }
+        self.sp -= 1;
+        self.ram[self.sp] = (val >> 8) as u8;
+        self.sp -= 1;
+        self.ram[self.sp] = val as u8;
+        Ok(())
+    }
+
+    pub fn push_reg(&mut self, reg: &str) -> EResult<()> {
+        self.push(self.reg[reg])
+    }
+
+    pub fn pop(&mut self) -> EResult<u16> {
+        if self.sp + 2 > self.ram.size() as u16 {
+            return Err("POP: No return address on the stack");
+        }
+        let low = self.ram[self.sp] as u16;
+        self.sp += 1;
+        let high = self.ram[self.sp] as u16;
+        self.sp += 1;
+        Ok((high << 8) | low)
+    }
+}
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    
+    #[test]
+    fn push_pop() {
+        let mut e = Emulator::new();
+
+        e.sp = 0x3fff;
+        e.push(0xabcd).expect("Push failed");
+        assert_eq!(e.sp, 0x3ffd);
+        assert_eq!(0xabcd, e.pop().expect("Fuck"));
+        assert_eq!(e.sp, 0x3fff);
+        assert_eq!(e.pop(), Err("POP: No return address on the stack"));
+
+        e.sp = 0x1;
+        assert_eq!(e.push(0x1234), Err("PUSH: No more stack space"));
+    }
+}
+
